feat #4 : 엔드포인트 경로 및 서비스 인터페이스 방식 적용
